requirements-dev.txt
pytest
pytest-asyncio
aiogram-tests
pytest-mock
